requirements.txt
flask
pandas
numpy
requests
sentence-transformers
faiss-cpu
